requirements.txt
seaborn
sqlite3
matplotlib
click
python-dateutil